[error] refactor errors
