+++ cita-executor/core/src/contracts/native/mod.rs
@@ -6,4 +6,4 @@ mod simple_storage;
 #[cfg(feature = "privatetx")]
 mod zk_privacy;
 
-pub use factory::{Contract, NativeError};
+pub use factory::Contract;
