+++ cita-executor/core/src/executed.rs
@@ -16,13 +16,14 @@
 
 //! Transaction execution format module.
 
-use std::fmt;
-
-use crate::types::errors::ReceiptError;
+use crate::types::errors::ExecutionError;
 use crate::types::log_entry::LogEntry;
 
-use cita_types::{Address, U256, U512};
+use cita_types::{Address, U256};
 use types::Bytes;
+
+/// Transaction execution result.
+pub type ExecutionResult = Result<Executed, ExecutionError>;
 
 /// Transaction execution receipt.
 #[derive(Debug)]
@@ -64,141 +65,4 @@ pub struct Executed {
     //pub state_diff: Option<StateDiff>,
     /// Transaction sender account nonce
     pub account_nonce: U256,
-}
-
-/// Result of executing the transaction.
-#[derive(PartialEq, Debug, Clone)]
-#[cfg_attr(feature = "ipc", binary)]
-pub enum ExecutionError {
-    /// Returned when there gas paid for transaction execution is
-    /// lower than base gas required.
-    NotEnoughBaseGas {
-        /// Absolute minimum gas required.
-        required: U256,
-        /// Gas provided.
-        got: U256,
-    },
-    /// Returned when block (gas_used + gas) > gas_limit.
-    ///
-    /// If gas =< gas_limit, upstream may try to execute the transaction
-    /// in next block.
-    BlockGasLimitReached {
-        /// Gas limit of block for transaction.
-        gas_limit: U256,
-        /// Gas used in block prior to transaction.
-        gas_used: U256,
-        /// Amount of gas in block.
-        gas: U256,
-    },
-    AccountGasLimitReached {
-        /// Account Gas limit left
-        gas_limit: U256,
-        /// Amount of gas in transaction
-        gas: U256,
-    },
-    /// Returned when transaction nonce does not match state nonce.
-    InvalidNonce {
-        /// Nonce expected.
-        expected: U256,
-        /// Nonce found.
-        got: U256,
-    },
-    /// Returned when cost of transaction (value + gas_price * gas) exceeds
-    /// current sender balance.
-    NotEnoughCash {
-        /// Minimum required balance.
-        required: U512,
-        /// Actual balance.
-        got: U512,
-    },
-    NoTransactionPermission,
-    NoContractPermission,
-    NoCallPermission,
-    /// Returned when internal evm error occurs.
-    ExecutionInternal(String),
-    /// Returned when generic transaction occurs
-    TransactionMalformed(String),
-}
-
-impl fmt::Display for ExecutionError {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        use self::ExecutionError::*;
-
-        let msg = match *self {
-            NotEnoughBaseGas { ref required, ref got } => format!("Not enough base quota. {} is required, but only {} paid", required, got),
-            BlockGasLimitReached {
-                ref gas_limit,
-                ref gas_used,
-                ref gas,
-            } => format!("Block quota limit reached. The limit is {}, {} has already been used, and {} more is required", gas_limit, gas_used, gas),
-            AccountGasLimitReached { ref gas_limit, ref gas } => format!("Account quota limit reached. The limit is {}, {} more is required", gas_limit, gas),
-            InvalidNonce { ref expected, ref got } => format!("Invalid transaction nonce: expected {}, found {}", expected, got),
-            NotEnoughCash { ref required, ref got } => format!("Cost of transaction exceeds sender balance. {} is required but the sender only has {}", required, got),
-            ExecutionInternal(ref msg) => msg.clone(),
-            TransactionMalformed(ref err) => format!("Malformed transaction: {}", err),
-            NoTransactionPermission => "No transaction permission".to_owned(),
-            NoContractPermission => "No contract permission".to_owned(),
-            NoCallPermission => "No call contract permission".to_owned(),
-        };
-
-        f.write_fmt(format_args!("Transaction execution error ({}).", msg))
-    }
-}
-
-/// Result of executing the transaction.
-#[derive(PartialEq, Debug, Clone)]
-#[cfg_attr(feature = "ipc", binary)]
-pub enum CallError {
-    /// Couldn't find the transaction in the chain.
-    TransactionNotFound,
-    /// Couldn't find requested block's state in the chain.
-    StatePruned,
-    /// Couldn't find an amount of gas that didn't result in an exception.
-    Exceptional,
-    /// Corrupt state.
-    StateCorrupt,
-    /// Error executing.
-    Execution(ExecutionError),
-}
-
-impl From<ExecutionError> for CallError {
-    fn from(error: ExecutionError) -> Self {
-        CallError::Execution(error)
-    }
-}
-
-impl fmt::Display for CallError {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        use self::CallError::*;
-
-        let msg = match *self {
-            TransactionNotFound => "Transaction couldn't be found in the chain".into(),
-            StatePruned => "Couldn't find the transaction block's state in the chain".into(),
-            Exceptional => "An exception happened in the execution".into(),
-            StateCorrupt => "Stored state found to be corrupted.".into(),
-            Execution(ref e) => format!("{}", e),
-        };
-
-        f.write_fmt(format_args!("Transaction execution error ({}).", msg))
-    }
-}
-
-/// Transaction execution result.
-pub type ExecutionResult = Result<Executed, ExecutionError>;
-
-impl From<ExecutionError> for ReceiptError {
-    fn from(error: ExecutionError) -> Self {
-        match error {
-            ExecutionError::NotEnoughBaseGas { .. } => ReceiptError::NotEnoughBaseQuota,
-            ExecutionError::BlockGasLimitReached { .. } => ReceiptError::BlockQuotaLimitReached,
-            ExecutionError::AccountGasLimitReached { .. } => ReceiptError::AccountQuotaLimitReached,
-            ExecutionError::InvalidNonce { .. } => ReceiptError::InvalidNonce,
-            ExecutionError::NotEnoughCash { .. } => ReceiptError::NotEnoughCash,
-            ExecutionError::NoTransactionPermission => ReceiptError::NoTransactionPermission,
-            ExecutionError::NoContractPermission => ReceiptError::NoContractPermission,
-            ExecutionError::NoCallPermission => ReceiptError::NoCallPermission,
-            ExecutionError::ExecutionInternal { .. } => ReceiptError::ExecutionInternal,
-            ExecutionError::TransactionMalformed { .. } => ReceiptError::TransactionMalformed,
-        }
-    }
 }
