+++ cita-executor/core/src/lib.rs
@@ -82,6 +82,7 @@ pub mod tests;
 
 pub mod authentication;
 pub mod contracts;
+pub mod exception;
 pub mod libexecutor;
 pub mod trie_db;
